+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [9.2.0] - 2024-08-11
+### Added
+* Add `ROBOTS_ALLOW_ALL_SHORT_URLS` config option.
+* Add `ROBOTS_USER_AGENTS` config option.
+* Add support for `laminas/laminas-servicemanager` v4.x.
+
+### Changed
+* Update to PHPUnit 11
+* Update to PHPStan 1.11
+
+### Deprecated
+* *Nothing*
+
+### Removed
+* *Nothing*
+
+### Fixed
+* *Nothing*
+
+
 ## [9.1.0] - 2024-04-14
 ### Added
 * Add `MEMORY_LIMIT` config option.